# Always build an IdModel during lowering