--- a/csrc/device_lower/id_model_options.h
+++ b/csrc/device_lower/id_model_options.h
@@ -16,7 +16,7 @@ namespace nvfuser {
 class IdModelOptions {
  public:
   IdModelOptions()
-      : build_id_model_(isOptionDisabled(DisableOption::IdModel)),
+      : build_id_model_(!isOptionDisabled(DisableOption::IdModel)),
         consumer_index_(
             isIdModelOptionEnabled(IdModelEnableOption::ConsumerIndex)),
         producer_index_(